[Estimate][Shared] - moved custom exceptions to shared project - created new exception when estimate is null - moved Estimate model to shared
