+++ EFDataAccessLibrary/Features/Estimates/EstimateProfile.cs
@@ -1,4 +1,6 @@
 ﻿using AutoMapper;
+using Shared.Estimates;
+
 
 namespace EFDataAccessLibrary.Features.Estimates
 {
