+++ EstimateForm.sln
@@ -5,7 +5,9 @@ VisualStudioVersion = 16.0.31019.35
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebAPI", "EstimateFormAPI-NotClean\WebAPI.csproj", "{28EEA741-3DCC-4989-9C92-41EA04EE873F}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EFDataAccessLibrary", "EFDataAccessLibrary\EFDataAccessLibrary.csproj", "{0CCCF4DA-6B61-4C3F-B962-B818982DA856}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EFDataAccessLibrary", "EFDataAccessLibrary\EFDataAccessLibrary.csproj", "{0CCCF4DA-6B61-4C3F-B962-B818982DA856}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shared", "Shared\Shared.csproj", "{D1F3B7FB-A1D4-4499-BF1D-DE7E9787CC2C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,6 +23,10 @@ Global
 		{0CCCF4DA-6B61-4C3F-B962-B818982DA856}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{0CCCF4DA-6B61-4C3F-B962-B818982DA856}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{0CCCF4DA-6B61-4C3F-B962-B818982DA856}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D1F3B7FB-A1D4-4499-BF1D-DE7E9787CC2C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D1F3B7FB-A1D4-4499-BF1D-DE7E9787CC2C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D1F3B7FB-A1D4-4499-BF1D-DE7E9787CC2C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D1F3B7FB-A1D4-4499-BF1D-DE7E9787CC2C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
