--- EstimateFormAPI-NotClean/Exceptions/EstimateNotComplete.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Runtime.Serialization;
-
-namespace WebAPI.Exceptions
-{
-    [Serializable]
-    public class EstimateNotComplete : Exception
-    {
-        protected EstimateNotComplete(SerializationInfo info, StreamingContext context) : base(info, context) { }
-
-        public EstimateNotComplete() : base("Provided request isn't complete (missing object parameters)") { }
-    }
-}
